added sam2 submodule and gui folder.
