+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "third_party/Depth-Anything-ONNX"]
 	path = third_party/Depth-Anything-ONNX
 	url = https://github.com/fabio-sim/Depth-Anything-ONNX.git
+[submodule "third_party/sam2"]
+	path = third_party/sam2
+	url = https://github.com/facebookresearch/sam2.git
