+++ data_engine/gui/__init__.py
@@ -1,0 +1,23 @@
+"""
+GUI module for SAM2 Data Engine
+Contains modularized GUI components
+"""
+
+from .center_panel import CenterPanel
+from .frame_viewer import FrameViewer
+from .left_panel import LeftControlPanel
+from .main_window import DataEngineMainWindow
+from .object_class_manager import ObjectClassManager
+from .right_panel import RightPanel
+from .workers import SAMProcessor, VideoProcessor
+
+__all__ = [
+    "DataEngineMainWindow",
+    "FrameViewer",
+    "ObjectClassManager",
+    "LeftControlPanel",
+    "RightPanel",
+    "CenterPanel",
+    "VideoProcessor",
+    "SAMProcessor",
+]
